+++ src/main/java/net/geforcemods/securitycraft/blocks/reinforced/ReinforcedCobwebBlock.java
@@ -2,7 +2,9 @@ package net.geforcemods.securitycraft.blocks.reinforced;
 
 import net.geforcemods.securitycraft.api.OwnableBlockEntity;
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
@@ -29,6 +31,13 @@ public class ReinforcedCobwebBlock extends BaseReinforcedBlock {
 		if (entity instanceof Player player && level.getBlockEntity(pos) instanceof OwnableBlockEntity be && be.isOwnedBy(player))
 			return;
 
-		entity.makeStuckInBlock(state, new Vec3(0.25D, 0.05D, 0.25D));
+		Vec3 slowness;
+
+		if (entity instanceof LivingEntity livingEntity && livingEntity.hasEffect(MobEffects.WEAVING))
+			slowness = new Vec3(0.5F, 0.25F, 0.5F);
+		else
+			slowness = new Vec3(0.25, 0.05F, 0.25);
+
+		entity.makeStuckInBlock(state, slowness);
 	}
 }